Add aws-sdk v3 dependencies
The development dependency was added to facilitate an upcoming aws-sdk
integration. The appraisals + circleci configuration were added mostly
blindly to parrot the other setups, assuming this will be needed for the
sake of continuous integration. The .ruby-version checked into the repo
points at Ruby 2.6 as of this writing, whereas the gemspec declares
support for Ruby >= 2.2.0, so it seems prudent for *someone* to be
testing these versions. Might as well be Circle.

We might technically try to include tests for aws-sdk v2 as well, since
v3 is supposed to be API-compatible [1]. I'm thinking that's probably
more effort than it's worth, though. Amazon is still apparently
supporting v2 of the gem [2], but it seems to largely be automated
maintenance (when they update the files that define the APIs). I'll
worry about developing against v3 first and foremost before even
entertaining the thought of backporting to v2.

References:

[1] https://github.com/aws/aws-sdk-ruby/blob/master/V3_UPGRADING_GUIDE.md

[2] https://github.com/aws/aws-sdk-ruby/commits/version-2