Use proper data type for session token credentials
While the STS stub data worked in the aws-3 appraisal, it was breaking
under the aws-2 appraisal with the error

  NoMethodError: undefined method `set?' for #<Aws::STS::Types::GetSessionTokenResponse:0x00007fcdf72d06f0>
  0: aws-sdk-core-2.11.388/lib/aws-sdk-core/plugins/request_signer.rb:110:in `missing_credentials?'
  1: aws-sdk-core-2.11.388/lib/aws-sdk-core/plugins/request_signer.rb:102:in `require_credentials'
  2: aws-sdk-core-2.11.388/lib/aws-sdk-core/plugins/s3_request_signer.rb:14:in `call'

Apparently some interface changed along the way so that the STS data
Just Works in v3. But a data type that works across both v2 *and* v3 is
the plain old Aws::Credentials class, which we can easily use to stub a
session token (just to make sure it appears in the spans when set).