Fix for encoded body in params
If the body has been encoded in some way (in this example using zlib)
then the body could contain a non UTF-8 string. Check when adding the
params for this and substitute a message for the missing body in this
case. Non UTF-8 strings will cause the JSON encoding that libhoney does
to send the events to break.