Safeguard #prettify against invalid characters
Per code review on #42.

We can still hex-escape the bytes that make up a string even if the
encoding is deemed invalid by Ruby. Since this (a) is handled by
redis-cli and (b) previously caused a hard error, it makes the most
sense to adapt the hex-escaping code to be resilient against bad
characters.