Fix `Redis.honeycomb_client = nil`
The idea of this attribute was that you'd be able to turn off the Redis
instrumentation by blanking out the Honeycomb client.

However, by using `||`, the code was *actually* falling back to the
global `Honeycomb.client` whenever we set `Redis.honeycomb_client =
nil`. The spec was passing erroneously because it was being run in a
context where `Honeycomb.client` was *also* `nil`. Oops!

By switching this to use the `defined?` check, we'll return `nil` from
`Redis.honeycomb_client` as long as it's been explicitly blanked out by
the user.