Add regression test for #65
The reworking of the Rails integration in #73 closes #65. We can reverse
engineer some minimum reproduction that *would* cause the behavior
described in the old code, but no longer does since we don't trigger
parameter parsing in the beeline. Again, all these issues are ultimately
caused by the parse failures, so avoiding parsing altogether prevents
the whole host of issues that people keep opening separately (but really
they're all the same).