Adding X-Forwarded-For to instrumented fields. (#91)
For any apps using the rack integration, the X-Forwarded-For header will now be automatically instrumented. This is useful for services that are behind proxies, where remote_addr will likely be the proxy IP.

X-Forwarded-For will be a single IP address or a list of IP addresses representing the originating IP address of a client connecting through an HTTP proxy or load balancer.

X-Forwarded-Proto is also captured, which will identify the protocol (HTTP or HTTPS) used by a client.

X-Forwarded-Port is sometimes used to identify the destination port that a client used to connect to the load balancer. It is used by AWS elastic load balancers amongst other load balancers and proxies.