Trigger a nightly build (#90)
Adds a workflow that runs once a day to lint and test whatever is on master. Followed these docs, no idea if it works or not https://circleci.com/docs/2.0/triggers/

It would be nice to stay on top of build failures like this https://github.com/honeycombio/beeline-ruby/pull/86 rather than have people submit PRs that fail to build because of some dependency failing.