remove children after sending (#98)
Resolves #84.

The spec changes look really bad, but the main change is to dump the existing specs into another `describe` block and add an additional spec to check that after the child is sent it won't try to send any sent children. Might be easiest to just look at the last two commits as the shuffling around is isolated to the first commit.