fix(client): reference current span in start_span (#105)
Updates the Client#start_span method to store a reference to the current span, rather than repeatedly referencing `context.current_span`.

The method `Span#send` is a public method, which means that it is possible for instrumentation to send the span from within a block passed to `Client#start_span`. Under those circumstances, one of two problematic scenarios can occur:

1) If the current span is _not_ the root span, then start_span will end up sending the parent span instead of the child span.
2) If the current span _is_ the root span, then start_span will raise `ArgumentError: no method name given`. This is due to the fact that `context.current_span` returns `nil` and `NilClass#send` expects a parameter, i.e. the method name to send.

By storing a reference to the current span, we can guarantee that `Client#start_span` will always be operating on the correct span. Given that `Span#send` is idempotent, calling `#send` after a block has already sent the span is safe.