Add Honeycomb.{with_field,with_trace_field} (#51)
The return value of `Honeycomb.add_field` (and similarly
`Honeycomb.add_field_to_trace`) is the `Libhoney::Event` object that had
the field added.

However, I'd often have the pattern in my code of using

```ruby
def something
  (some || sort && of ** expression).tap do |value|
    Honeycomb.add_field('field', value)
  end
end
```

That is, I still want the return value of the `#something` method, but I
wanted to track its value in a Honeycomb field. The methods added in
this commit allow the more natural pattern of wrapping the computation
in a block like so:

```ruby
def something
  Honeycomb.with_field('field') do
    some || sort && of ** expression
  end
end
```

It's just a minor feature, but one that would make some of my code a
little nicer looking. Especially with predicates, where I currently have
code like `question?.tap { ... }`.

Co-authored-by: Mike Goldsmith <goldsmith.mike@gmail.com>