Add test case for 5xx response (#132)
* adding back test coverage introduced in https://github.com/honeycombio/beeline-ruby/pull/123
* #123 was reverted so we can continue to release a non-major version off main

It has become expected that `error` and `error_detail` will be present on events from Rails when
HTTP errors occur. The change that was reverted (#123), fixed a problem with status codes, but
also resulted in these error fields disappearing from events. These tests are being added to
test for the current error field behavior, so that the next fix for status code inclusion also keeps
these error fields.