Improve performance of Redis command serialization (#146)
Fixes #57.

There was a major performance bottleneck in the Redis integration when
prettifying the `redis.command` field. The original implementation would
scan the string character-by-character and manually push
possibly-escaped results to another string. This implementation replaces
that manual logic with some strategic uses of `String#encode!` and
`String#gsub!`

There's a minor regression in the behavior with the interaction between
invalid UTF-8 bytes and basic backslash escaping, but I think the payoff
is well worth it. I added tests to codify this new behavior.