Allow backtrace to be sent with errors (#160)
 often find myself looking for errors and finding something like:

NoMethodError: No method `split' on <String#0x00012121212>
This is decidedly unhelpful for figuring out where exactly this happened, especially if the specific code isn't instrumented because normally it's trivial. This means the error only bubbles up and exists on the next-nearest parent or, even worse, the root span. By allowing users to configure and provide a limited set of the backtrace, they can get at least the chance for context without finding the error'ed event in Honeycomb then trying to find the same error in their logs to get the backtrace of the location.

While the limited subset may still mean they have to do that, this at least opens the door to providing just enough context to find where an error is coming from.