add specs for default handling
The existing specs all fiddled with the event handlers in some way.
This adds a context and specs for what we expect the default handler to
do when the Beeline is subscribed to an event type.

Testing Rails gets wordy, so I've opted to turn off the line length cop
for this spec file instead of increasing the noisiness of the test code
with sectional waivers.