add a test that event data exists
It is possible to arrange event tests such that the event_data used for
the shared examples is an empty array. For example, if the Beeline's
Honeycomb::Client does not really get instantiated by a let(:client)
statement because "client" is never used in a test. The exclamation
let!() is easy to miss or forget. So this spec asserts that the examples
have at least one thing in the event_data collection to test against.