normalize ActiveSupport::Notification exceptions into error / error_d… (#166)
* add specs for default and custom handling

The existing specs all fiddled with the event handlers in some way.
This adds a context and specs for what we expect the default handler to
do when the Beeline is subscribed to an event type.

* turn off line length cop for this spec

Testing Rails gets wordy, so we opted to turn off the line length cop
for this spec file instead of increasing the noisiness of the test code
with sectional waivers.

* add a test that event data exists

It is possible to arrange event tests such that the event_data used for
the shared examples is an empty array. For example, if the Beeline's
Honeycomb::Client does not really get instantiated by a let(:client)
statement because "client" is never used in a test. The exclamation
let!() is easy to miss or forget. So this spec asserts that the examples
have at least one thing in the event_data collection to test against.

* use the 2-element exception info

We've still got Rails 4.2 in the test matrix and :exception_object wasn't
added until Rails 5.

Co-authored-by: Robb Kidd <robbkidd@honeycomb.io>