fix: clean invalid UTF8 from trace field values when generating propagation header (#232)
## Which problem is this PR solving?

- Resolves #231 

## Short description of the changes

### Tests to Demonstrate the Problem

* Tests added to cover invalid UTF8 characters in trace field values for
classic and E&S modes of Honeycomb propagation

### MarshalTraceContext

* Replace invalid UTF8 characters when serializing in Honeycomb-flavored
propagators

Mixes in the data cleaning logic already present in Libhoney::Cleaner to
prepare trace data for JSON serialization.

* Refactor the encoding logic for serialization in an extracted method
to avoid having the two implementations diverge.

### UnmarshalTraceContext

* Decode trace context with the decoder that matches the encoder.

Testing revealed that decoding trace context would fail to produce a
string parseable as JSON if some UTF8 characters were included in the
data, even if those characters are valid. Because we encode with
urlsafe, we need to decode with urlsafe.

### Linting

* Disable the Style/AccessModifierDeclarations Rubocop rule.

Sometimes we group methods under "private", sometimes we inline a symbol
reference to a method with "module_method". We remain flexible. This
rule does not serve us and we do not serve it.